requirements.txt
absl-py==2.1.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
audio-separator==0.17.4
audioread==3.0.1
beartype==0.18.5
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==7.1.2
colorama==0.4.6
commonmark==0.9.1
contextlib2==21.6.0
Cython==3.0.10
decorator==4.4.2
diffq==0.2.4
einops==0.8.0
exceptiongroup==1.2.1
ffmpeg-python==0.2.0
filelock==3.15.4
flatbuffers==1.12
fsspec==2024.6.1
future==1.0.0
gast==0.4.0
google-auth==2.30.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.64.1
h11==0.12.0
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
huepy==1.2.1
hyperframe==6.0.1
idna==3.7
imageio==2.35.1
imageio-ffmpeg==0.5.1
iniconfig==2.0.0
instabot==0.117.0
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
julius==0.2.7
keras==2.9.0
Keras-Preprocessing==1.1.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.43.0
Markdown==3.6
MarkupSafe==2.1.5
mkl==2021.4.0
ml-collections==0.1.1
mock==5.1.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.8
mutagen==1.47.0
networkx==3.3
norbert==0.2.1
numba==0.60.0
numpy==2.0.0
oauthlib==3.2.2
onnx==1.16.1
onnx2torch==1.5.14
opt-einsum==3.3.0
packaging==24.1
pandas==1.5.3
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
proglog==0.1.10
protobuf==5.27.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydub==0.25.1
Pygments==2.18.0
pyperclip==1.9.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
resampy==0.4.3
responses==0.25.3
rfc3986==1.5.0
rich==12.6.0
rotary-embedding-torch==0.6.2
rsa==4.9
samplerate==0.1.0
schedule==1.2.2
scikit-learn==1.5.0
scipy==1.14.0
sentry-sdk==1.45.1
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spleeter==2.4.0
spotify_dl==8.9.0
spotipy==2.24.0
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.3.2
typing_extensions==4.12.2
urllib3==1.26.19
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
youtube-dl==2021.12.17
yt-dlp==2023.12.30
